% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_epsilon}
\alias{add_epsilon}
\title{Expand grid}
\usage{
add_epsilon(G, epsilon, by, subset = colnames(G))
}
\arguments{
\item{G}{matrix with one parameter setting in each line}

\item{epsilon}{width of interval to be added around each points in G}

\item{by}{distance between points to be added within intervals around points in G}

\item{subset}{subset of parameters for which grid points should be added}
}
\value{
object with the same columns as G
}
\description{
Expand and refine parameter grid for simulation
}
\author{
Florian Klinglmueller
}

