% Generated by roxygen2 (4.0.0): do not edit by hand
\name{casesim}
\alias{casesim}
\title{Case study simulations "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^6$ simulated trials.}
\description{
A dataset containing the very simulation results used
shown in the article "Estimation after blinded interim
analyses". The simulation may be re-run using code shown in the
vignette "simulations".
}

