\name{gridsim}
\alias{gridsim}
\title{Simulation results from "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^6$ simulated trials.}
\description{
A dataset containing the very simulation results used shown
in the article "Estimation after blinded interim analyses".
The simulation may be re-run using code shown in the
\code{\link{vignette("SimulateBias")}}.
}

