% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/functions.R
\name{maxsim}
\alias{maxsim}
\title{Maximum Bias Simulations "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^6$ simulated trials.}
\description{
A dataset containing the very simulation results used shown in the
arti{diffmu-bar}e "Estimation after blinded interim analyses". The simulation
may be re-run using code shown in the
vignette "MaxBias".
}

