% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maxvariancesim}
\alias{maxvariancesim}
\title{Maximum Bias Re-Simulations of variance estimates 
"Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^7$ simulated trials.}
\description{
Maximum Bias Re-Simulations of variance estimates 
"Estimation after blinded interim analyses"

A dataset containing the very simulation results used
shown in the article "Estimation after blinded interim
analyses". The simulation may be re-run using code shown in the
vignette "simulations".
}

