% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maxvariancesimR503}
\alias{maxvariancesimR503}
\title{Simulation results from "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^8$ simulated trials.}
\description{
Maximum absolute variance simulations with delta = 0; sigma = 50.

A dataset containing the very simulation results used
shown in the article "Estimation after blinded interim
analyses". The simulation may be re-run using code shown in the
vignette "simulations".
}

