% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maxvariancesimRL3}
\alias{maxvariancesimRL3}
\title{Simulation results from "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^8$ simulated trials.}
\description{
Simulation results from "Estimation after blinded interim analyses"

A dataset containing the very simulation results used
shown in the article "Estimation after blinded interim
analyses". The simulation may be re-run using code shown in the
vignette "simulations".
}

