% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mean_bias_plot}
\alias{mean_bias_plot}
\title{Mean bias plot}
\usage{
mean_bias_plot(gridsim, thmax_bias = TRUE)
}
\arguments{
\item{gridsim}{Simulation results}

\item{thmax_bias}{Should the theoretical upper bound be plotted (needs to be in the data)}
}
\value{
ggplot object
}
\description{
Plot the bias of the effect estimate following blinded sample size reassessment
}
\author{
float
}

