% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{n2r}
\alias{n2r}
\title{Sample size reassessment rules}
\usage{
n2r(S1OS, delta0, n1, adj = FALSE, n2min = 0, n2max = Inf, cf = 1,
  alpha = 0.025, beta = 0.2)
}
\arguments{
\item{S1OS}{interim estimate of the standard deviation}

\item{delta0}{assumed effect size}

\item{n1}{first-stage sample size}

\item{adj}{should the adjusted rule be used}

\item{n2min}{minimum second-stage sample size}

\item{n2max}{maximum second-stage sample size}

\item{cf}{correction facton}

\item{alpha}{significance level (one-sided)}

\item{beta}{targeted Type II error rate}
}
\value{
adjusted second-stage sample size
}
\examples{
## Example from Kieser et al. (2003)
n1 = 15
S1OS = 6
delta0 = 5.5
n2r(S1OS,delta0,n1)
n2r(S1OS,delta0,n1,adj=T)
}
\author{
Florian Klinglmueller
}

