% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Coverage plot}
\usage{
plot_coverage(gridsim, adjusted = F)
}
\arguments{
\item{gridsim}{Simulation results}

\item{adjusted}{should the results for the adjusted interim variance estimate be plotted}
}
\value{
ggplot object
}
\description{
Plot coverage probabilities of confidence intervals following blinded sample size reassessment
}
\author{
float
}

