% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsim.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot parameter grids}
\usage{
plot_grid(maxsim, G2s, G2e, what)
}
\arguments{
\item{maxsim}{Results of previous simulation}

\item{G2s}{Parameter values at which optimum occured}

\item{G2e}{Refined parameter grid}
}
\description{
Plot to compare different grids of parameter values
}
\author{
Florian Klinglmueller
}

