% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{restsim}
\alias{restsim}
\title{Simulation results from "Estimation after blinded interim analyses"}
\format{A \code{data.frame}, each line corresponds to the outcome from
one simulation run with $10^7$ simulated trials.}
\description{
Sample size rules with second stage sample size limited \code{n2max=2*n1}

A dataset containing the very simulation results used
shown in the article "Estimation after blinded interim
analyses". The simulation may be re-run using code shown in the
vignette "simulations".
}

