% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxsim.R
\name{select_results}
\alias{select_results}
\title{Select simulation results}
\usage{
select_results(maxsim, what, base_columns = c("delta", "sigma", "d", "n1",
  "s"), functional = which.min)
}
\arguments{
\item{maxsim}{\code{\link{data.table}} of simulation results}

\item{what}{character vector of column names according to which the selection should be performed}

\item{base_columns}{columns to keep from original data.table use \code{NULL} to keep all columns}

\item{functional}{function applied columns \code{what} whithin each unique first stage sample size that returns an index of the selected rows}
}
\value{
\code{data.table} with selected rows, plus one row giving the column name the selection was based on
}
\description{
Select something for each first stages sample size from a \code{\link{data.table}} of simulation results
}
\examples{
data(maxsim)
## select largest (absolute) mean.bias for each first stage sample size
select_results(maxsim,'mean.bias')
}
\author{
Florian Klinglmueller
}

