% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sem_plot}
\alias{sem_plot}
\title{SEM plot}
\usage{
sem_plot(gridsim, adjusted = F)
}
\arguments{
\item{gridsim}{Simulation results}

\item{adjusted}{should the results for the adjusted interim variance estimate be plotted}
}
\value{
ggplot object
}
\description{
Plot the variance of the effect estimate (the square of the standard error of the mean) following blinded sample size reassessment
}
\author{
float
}

