% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simulate_batch}
\alias{simulate_batch}
\title{Batch simulate blinded sample size review bias}
\usage{
simulate_batch(G, runs)
}
\arguments{
\item{G}{Object with paramter settings for \code{delta}, \code{sigma} and \code{n1} in each row}

\item{runs}{Number of simulation runs}
}
\value{
Object with parameters and simulation results in each row
}
\description{
Simulate bias of blinded sample size review for a range of parameter settings
}
\examples{
data(maxsim)
G2s <- select_results(maxsim,c('mean.bias'))
G2e <- add_epsilon(G2s,.05,.05,subset=c('delta','sigma'))
\\donotrun{plot_grid(maxsim,G2s,G2e,what='delta')}
\\donotrun{resim1 <- simulate_batch(G2e,10^3)}
}
\author{
Florian Klinglmueller
}

