% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{variance_bias_plot}
\alias{variance_bias_plot}
\title{Variance bias plot}
\usage{
variance_bias_plot(gridsim)
}
\arguments{
\item{gridsim}{Simulation results}
}
\value{
ggplot object
}
\description{
Plot the bias of the variance of the effect estimate following blinded sample size reassessment
}
\author{
float
}

