% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_simVBIA}
\alias{combine_simVBIA}
\title{Combine simulation results}
\usage{
combine_simVBIA(...)
}
\arguments{
\item{...}{any number of simulation results from \code{simVBIA} with \code{fulldata} set to \code{TRUE}}
}
\value{
a simulation result object
}
\description{
Combine the fulldata results of two simulation runs
}
\author{
float
}

