% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coveragesim.R
\name{coverage_sim}
\alias{coverage_sim}
\title{coverage sim}
\usage{
coverage_sim(search = 10^4, resim = 10^6, use_mclapply2 = TRUE)
}
\arguments{
\item{search}{number of simulation runs for the wide search}

\item{resim}{number of simulation runs for the resimulation of found maxima}

\item{use_mclapply2}{use mclapply2 in simulations}
}
\value{
dataframe with simulation results
}
\description{
run coverage sims
}
\author{
float
}

