% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragesim.R
\name{coverage_sim}
\alias{coverage_sim}
\title{coverage sim}
\usage{
coverage_sim(base = maxsim, search = 10^4, resim = 10^6,
  use_mclapply2 = TRUE, mem_factor = 16, epsilon = 0.18, by = 0.06)
}
\arguments{
\item{base}{initial simulation across a grid to base the search on}

\item{search}{number of simulation runs for the wide search}

\item{resim}{number of simulation runs for the resimulation of found maxima}

\item{use_mclapply2}{use mclapply2 in simulations}

\item{mem_factor}{number of cores that can handle 10^7 simulation runs, will be used to automatically determine the number of cores}

\item{epsilon}{width of parameter grid around initial values}

\item{by}{fineness of grid around initial values}
}
\value{
dataframe with simulation results
}
\description{
run coverage sims
}
\author{
float
}

