% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxsim.R
\name{simulate_batch}
\alias{simulate_batch}
\title{Batch simulate blinded sample size review bias}
\usage{
simulate_batch(G, runs, use_mclapply2 = FALSE, multicore = TRUE)
}
\arguments{
\item{G}{Object with paramter settings for \code{delta}, \code{sigma} and \code{n1} in each row}

\item{runs}{Number of simulation runs}

\item{use_mclapply2}{use mclapply2 from package bt88.03.704 which implements a status bar (see details)}
}
\value{
Object with parameters and simulation results in each row
}
\description{
Simulate bias of blinded sample size review for a range of parameter settings
}
\details{
To use mclapply2 from bt88.03.704 you have to install this library (e.g. \code{devtools::install_github('floatofmath/bt88.03.704')}) unfortunately the statusbar implementation does not work with Rstudio.
}
\examples{
data(maxsim)
G2s <- select_results(maxsim,c('mean.bias'))
G2e <- add_epsilon(G2s,.05,.05,subset=c('delta','sigma'))
\\donotrun{plot_grid(maxsim,G2s,G2e,what='delta')}
\\donotrun{resim1 <- simulate_batch(G2e,10^3)}

}
\author{
Florian Klinglmueller
}

